requirements.txt
telethon
spotipy
discord.py
pyGithub
pytz
Pillow